First pass at converting vocab editor to a Typescript SFC Vue app
Still need to revise the Ajax functionality so it uses Axios instead
of jQuery calls via Play's JS router. Also lots of type annotations
to add.
