+++ modules/admin/app/assets/js/vocabeditor/components/_concept-creator.vue
@@ -1,0 +1,69 @@
+<script lang="ts">
+
+import ConceptDataEditor from './_concept-data-editor';
+import VocabEditorApi from "../api";
+
+export default {
+  components: {ConceptDataEditor},
+  props: {
+    api: VocabEditorApi,
+    lang: String,
+    langData: Object,
+    localeHelpers: Object,
+    data: Object,
+  },
+  data: function () {
+    return {
+      loading: false,
+      saving: false,
+      saved: false,
+      error: false,
+      errors: {},
+      state: this.data,
+    }
+  },
+  methods: {
+    createItem: function (toSave) {
+      this.loading = true;
+      this.saving = true;
+      this.api.createItem(toSave)
+          .then(item => this.$emit('item-data-saved', item))
+          .then(() => {
+            this.saved = true;
+            this.error = false;
+            this.saving = false;
+            this.loading = false;
+          }).fail(err => {
+        this.error = true;
+        this.saving = false;
+        this.loading = false;
+        if (err.status === 400 && err.responseJSON) {
+          this.errors = err.responseJSON;
+        }
+      });
+    },
+  },
+}
+</script>
+
+<template>
+  <div id="concept-creator" class="form-horizontal">
+    <h3 id="concept-creator-title">New Concept</h3>
+    <concept-data-editor
+        v-bind:lang="lang"
+        v-bind:id="null"
+        v-bind:create="true"
+        v-bind:data="state"
+        v-bind:langData="langData"
+        v-bind:locale-helpers="localeHelpers"
+        v-bind:dirty="true"
+        v-bind:loading="loading"
+        v-bind:saving="saving"
+        v-bind:saved="saved"
+        v-bind:error="error"
+        v-bind:errors="errors"
+        v-on:item-data-saved="createItem"
+        v-on:cancel-create="$emit('cancel-create')"
+    />
+  </div>
+</template>
